requirements.txt
pillow
djangorestframework
django-cors-headers
djangorestframework-simplejwt
django-allauth 
django-rest-auth